# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KLSI 4.0 API",
    "description": "Assessment engine and analytics API for KLSI 4.0",
    "version": "4.0.0"
  },
  "paths": {
...
    "/reports/self": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Self Reports",
        "operationId": "list_self_reports_reports_self_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportSummaryPayload"
                  },
                  "type": "array",
                  "title": "Response List Self Reports Reports Self Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/reports/{session_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report",
        "operationId": "get_report_reports__session_id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IndividualReportPayload"
                    },
                    {
                      "$ref": "#/components/schemas/TeamReportPayload"
                    }
                  ],
                  "title": "Response Get Report Reports  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{session_id}/share": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Create Report Share",
        "operationId": "create_report_share_reports__session_id__share_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportShareCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportShareOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/shared/{share_token}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Shared Report",
        "operationId": "get_shared_report_reports_shared__share_token__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "share_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IndividualReportPayload"
                    },
                    {
                      "$ref": "#/components/schemas/TeamReportPayload"
                    }
                  ],
                  "title": "Response Get Shared Report Reports Shared  Share Token  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IndividualReportPayload": {
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "title": "Sessionid"
          },
          "generatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedat"
          },
          "kind": {
            "type": "string",
            "enum": [
              "individual"
            ],
            "const": "individual",
            "title": "Kind",
            "default": "individual"
          },
          "raw": {
            "type": "object",
            "title": "Raw"
          },
          "percentiles": {
            "type": "object",
            "title": "Percentiles"
          },
          "style": {
            "type": "object",
            "title": "Style"
          },
          "lfi": {
            "type": "object",
            "title": "Lfi"
          },
          "analytics": {
            "type": "object",
            "title": "Analytics"
          },
          "visualization": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visualization"
          },
          "sessionDesigns": {
            "anyOf": [
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessiondesigns"
          },
          "learningSpace": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learningspace"
          },
          "enhancedAnalytics": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enhancedanalytics"
          },
          "notes": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "owner": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "shareContext": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportShareContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "responsibleUseNotice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsibleusenotice"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "raw",
          "percentiles",
          "style",
          "lfi",
          "analytics"
        ],
        "title": "IndividualReportPayload"
      },
...
      "ReportDialecticSummary": {
        "properties": {
          "acce": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acce"
          },
          "aero": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aero"
          },
          "intensity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intensity"
          }
        },
        "type": "object",
        "title": "ReportDialecticSummary"
      },
      "ReportFlexibilitySummary": {
        "properties": {
          "lfiScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lfiscore"
          },
          "percentile": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentile"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "levelLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Levellabel"
          }
        },
        "type": "object",
        "title": "ReportFlexibilitySummary"
      },
      "ReportLongitudinalSummary": {
        "properties": {
          "previousSessionId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previoussessionid"
          },
          "previousSessionDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previoussessiondate"
          },
          "timeElapsedDays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeelapseddays"
          },
          "deltaAcce": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deltaacce"
          },
          "deltaAero": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deltaaero"
          },
          "deltaLfi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deltalfi"
          },
          "deltaIntensity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deltaintensity"
          }
        },
        "type": "object",
        "title": "ReportLongitudinalSummary"
      },
      "ReportShareContext": {
        "properties": {
          "shareId": {
            "type": "integer",
            "title": "Shareid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "title": "Sessionid"
          },
          "mediatorEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediatoremail"
          },
          "mediatorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediatorname"
          },
          "ownerName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ownername"
          },
          "ownerEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owneremail"
          },
          "expiresAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "shareId",
          "sessionId"
        ],
        "title": "ReportShareContext"
      },
      "ReportShareCreate": {
        "properties": {
          "mediatorEmail": {
            "type": "string",
            "format": "email",
            "title": "Mediatoremail",
            "description": "Email mediator yang diberi akses"
          },
          "expiresInHours": {
            "type": "integer",
            "maximum": 336.0,
            "minimum": 1.0,
            "title": "Expiresinhours",
            "description": "Durasi link berlaku dalam jam (maks 14 hari)",
            "default": 72
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "mediatorEmail"
        ],
        "title": "ReportShareCreate"
      },
      "ReportShareOut": {
        "properties": {
          "shareId": {
            "type": "integer",
            "title": "Shareid"
          },
          "sessionId": {
            "type": "integer",
            "title": "Sessionid"
          },
          "mediatorEmail": {
            "type": "string",
            "format": "email",
            "title": "Mediatoremail"
          },
          "mediatorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediatorname"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "title": "Expiresat"
          },
          "shareToken": {
            "type": "string",
            "title": "Sharetoken"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "shareId",
          "sessionId",
          "mediatorEmail",
          "mediatorName",
          "expiresAt",
          "shareToken",
          "note"
        ],
        "title": "ReportShareOut"
      },
      "ReportStyleSummary": {
        "properties": {
          "styleCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stylecode"
          },
          "styleName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stylename"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "quadrant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quadrant"
          }
        },
        "type": "object",
        "title": "ReportStyleSummary"
      },
      "ReportSummaryPayload": {
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "title": "Sessionid"
          },
          "generatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedat"
          },
          "learningStyle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportStyleSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "nineStyle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportStyleSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "flexibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportFlexibilitySummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "dialectic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportDialecticSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "longitudinal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportLongitudinalSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "sessionId"
        ],
        "title": "ReportSummaryPayload"
      },
...
      "TeamReportPayload": {
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "title": "Sessionid"
          },
          "generatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedat"
          },
          "kind": {
            "type": "string",
            "enum": [
              "team"
            ],
            "const": "team",
            "title": "Kind",
            "default": "team"
          },
          "teamId": {
            "type": "integer",
            "title": "Teamid"
          },
          "analytics": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analytics"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "teamId"
        ],
        "title": "TeamReportPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login or /auth/register"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}